refactor(layering): extract RuntimeStatus into a runtime-status crate
The component status tracker had no dependency on any runtime internals, yet it
sat inside the runtime crate's single codegen unit. Move it to a below-runtime
crate (foundation tier) so it compiles in parallel with other runtime-tier work
and so components beneath the orchestrator can report and await status without
depending on runtime.

runtime::status re-exports the tracker, leaving all call sites unchanged. The
status-by-name update method becomes pub for the cluster scheduler, which sets a
component status the typed helpers don't cover.
