+++ layers.toml
@@ -182,6 +182,7 @@ runtime-parameter-spec = "foundation"
 runtime-proto = "foundation"
 runtime-query-engine = "foundation"
 runtime-rate-control = "foundation"
+runtime-status = "foundation"
 s3_vectors = "foundation"
 s3_vectors_metadata_filter = "foundation"
 scheduler = "foundation"
